+++ packages/browser-wallet/src/popup/pages/Account/TransactionLog/i18n/en.ts
@@ -6,6 +6,7 @@ const t = {
     blockHash: 'Block hash',
     events: 'Events',
     rejectReason: 'Reject reason',
+    noTransactions: 'This account has no transactions yet.',
 };
 
 export default t;